Restructured the WebSocket namespace
